Integrate Eddie's comments and Yoni's refactoring suggestion.
