+++ hbase-client/src/main/java/org/apache/omid/transaction/TableAccessWrapper.java
@@ -24,6 +24,7 @@ import org.apache.hadoop.hbase.client.Result;
 import org.apache.hadoop.hbase.client.Put;
 import org.apache.hadoop.hbase.client.Get;
 
+//This interface is used to wrap the HTableInterface and Region object when doing client and server side filtering accordingly.
 public interface TableAccessWrapper {
 
     public Result[] get(List<Get> get) throws IOException;
