+++ hbase-coprocessor/src/main/java/org/apache/omid/transaction/OmidSnapshotFilter.java
@@ -60,7 +60,7 @@ public class OmidSnapshotFilter extends BaseRegionObserver {
     @VisibleForTesting
     private CommitTable.Client commitTableClient;
 
-    private SnapshotFilter snapshotFilter;
+    private SnapshotFilterImpl snapshotFilter;
 
     final static String OMID_SNAPSHOT_FILTER_CF_FLAG = "OMID_SNAPSHOT_FILTER_ENABLED";
 
@@ -79,7 +79,7 @@ public class OmidSnapshotFilter extends BaseRegionObserver {
         }
         commitTableClient = initAndGetCommitTableClient();
         
-        snapshotFilter = new SnapshotFilter(commitTableClient);
+        snapshotFilter = new SnapshotFilterImpl(commitTableClient);
         
         LOG.info("Snapshot filter started");
     }
